refactor: 시간 정보가 Long으로 저장되도록 변경

+++ server/src/main/java/wap/starlist/domain/Bookmark.java
@@ -24,11 +24,11 @@ public class Bookmark {
 
     private String url;
 
-    private Integer dateAdded;
+    private Long dateAdded;
 
-    private Integer dateGroupModified;
+    private Long dateGroupModified;
 
-    private Integer dateLastUsed;
+    private Long dateLastUsed;
 
     private Integer position;
